refactor(sysinfo): extract DashboardService to internal/sysinfo
