+++ internal/sysinfo/dashboard_service_test.go
@@ -1,8 +1,8 @@
-// file: internal/server/dashboard_service_test.go
+// file: internal/sysinfo/dashboard_service_test.go
 // version: 1.1.0
 // guid: a0b1c2d3-e4f5-6a7b-8c9d-0e1f2a3b4c5d
 
-package server
+package sysinfo
 
 import (
 	"testing"
